+++ test/libsolidity/syntaxTests/array/calldata_assign.sol
@@ -1,8 +1,7 @@
 pragma experimental ABIEncoderV2;
 contract Test {
-    struct S { int a; }
-    function f(S calldata s) external { s.a = 4; }
+    function f(uint256[] calldata s) external { s[0] = 4; }
 }
 // ----
 // Warning: (0-33): Experimental features are turned on. Do not use experimental features on live deployments.
-// TypeError: (114-117): Calldata structs are read-only.
+// TypeError: (98-102): Calldata arrays are read-only.