+++ test/libsolidity/syntaxTests/structs/calldata_array_assign.sol
@@ -7,4 +7,4 @@ contract Test {
 }
 // ----
 // Warning: (0-33): Experimental features are turned on. Do not use experimental features on live deployments.
-// TypeError: (144-147): Expression has to be an lvalue.
+// TypeError: (144-147): Calldata structs are read-only.