+++ test/libsolidity/syntaxTests/structs/memory_to_calldata.sol
@@ -6,7 +6,7 @@ contract Test {
 }
 // ----
 // Warning: (0-33): Experimental features are turned on. Do not use experimental features on live deployments.
-// TypeError: (114-115): Expression has to be an lvalue.
+// TypeError: (114-115): External function arguments of reference type are read-only.
 // TypeError: (118-122): Type struct Test.S memory is not implicitly convertible to expected type struct Test.S calldata.
-// TypeError: (178-179): Expression has to be an lvalue.
+// TypeError: (178-179): External function arguments of reference type are read-only.
 // TypeError: (182-183): Type struct Test.S memory is not implicitly convertible to expected type struct Test.S calldata.